+++ app/src/main/java/com/example/teumteum/ui/main/HomeFragment.kt
@@ -20,6 +20,7 @@ import com.example.teumteum.data.entities.TodoHomeItem
 import com.example.teumteum.data.entities.WishItem
 import com.example.teumteum.ui.calendar.CalendarVPAdapter
 import com.example.teumteum.ui.filling.FillingActivity01Fragment
+import com.example.teumteum.ui.todo.TodoRegisterFragment
 import com.example.teumteum.ui.wish.WishlistFragment
 import com.google.android.material.bottomnavigation.BottomNavigationView
 
@@ -82,7 +83,7 @@ class HomeFragment : Fragment(), IDateClickListener {
         }
 
         binding.fabAddIv.setOnClickListener {
-            val bottomSheet = BottomSheetRegisterFragment()
+            val bottomSheet = TodoRegisterFragment()
             bottomSheet.show(parentFragmentManager, bottomSheet.tag)
         }
