+++ app/src/main/java/com/example/teumteum/ui/wish/WishRegisterFragment.kt
@@ -1,5 +1,6 @@
 package com.example.teumteum.ui.wish
 
+import android.app.Dialog
 import android.content.res.ColorStateList
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,13 +9,19 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.example.teumteum.R
 import com.example.teumteum.databinding.FragmentWishRegisterBinding
+import com.example.teumteum.ui.todo.TodoRegisterFragment
+import com.google.android.material.bottomsheet.BottomSheetBehavior
+import com.google.android.material.bottomsheet.BottomSheetDialog
+import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import com.google.android.material.button.MaterialButton
 
-class WishRegisterFragment : Fragment() {
+class WishRegisterFragment : BottomSheetDialogFragment() {
 
     private lateinit var binding: FragmentWishRegisterBinding
     private var selectedTimeButton: View? = null
     private var selectedCategoryButton: View? = null
+
+    private var isWishSelected = true
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -27,6 +34,18 @@ class WishRegisterFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+
+        binding.btnTodo.setOnClickListener {
+            if (isWishSelected) {
+                binding.btnTodo.setImageResource(R.drawable.btn_todo_active_sv)
+                binding.btnWish.setImageResource(R.drawable.btn_wish_deactive_sv)
+                isWishSelected = false
+
+                childFragmentManager.beginTransaction()
+                    .replace(R.id.register_fragment_container, TodoRegisterFragment())
+                    .commit()
+            }
+        }
 
         setupTimeButtons()
         setupCategoryButtons()
@@ -90,5 +109,37 @@ class WishRegisterFragment : Fragment() {
                 selectedCategoryButton = button
             }
         }
+    }
+
+    override fun onStart() {
+        super.onStart()
+
+        dialog?.let { dialog ->
+            val bottomSheet = dialog.findViewById<View>(com.google.android.material.R.id.design_bottom_sheet)
+            bottomSheet?.let {
+                val screenHeight = resources.displayMetrics.heightPixels
+                val desiredHeight = (screenHeight * 0.84).toInt()
+
+                it.layoutParams.height = desiredHeight
+                it.requestLayout()
+
+                val behavior = BottomSheetBehavior.from(it)
+                behavior.peekHeight = desiredHeight
+                behavior.state = BottomSheetBehavior.STATE_COLLAPSED
+                behavior.isDraggable = false // 확장 불가능
+            }
+        }
+    }
+
+    override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
+        val dialog = super.onCreateDialog(savedInstanceState) as BottomSheetDialog
+
+        dialog.setOnShowListener { dialogInterface ->
+            val bottomSheet = (dialogInterface as BottomSheetDialog)
+                .findViewById<View>(com.google.android.material.R.id.design_bottom_sheet)
+            bottomSheet?.setBackgroundResource(R.drawable.calendar_background)
+        }
+
+        return dialog
     }
 }
