--- app/src/main/res/layout/fragment_bottom_sheet_register.xml
@@ -1,53 +1,0 @@
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/bottom_sheet_root"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <androidx.core.widget.NestedScrollView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:fillViewport="true">
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical"
-            android:padding="17dp"
-            android:background="@drawable/calendar_background">
-
-            <View
-                android:layout_width="54dp"
-                android:layout_height="4dp"
-                android:layout_gravity="center_horizontal"
-                android:layout_marginBottom="15dp"
-                android:background="@drawable/bg_bottom_sheet_handle" />
-
-            <LinearLayout
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:orientation="horizontal">
-
-                <ImageView
-                    android:id="@+id/btn_todo"
-                    android:layout_width="60dp"
-                    android:layout_height="28dp"
-                    android:layout_marginEnd="8dp"
-                    android:src="@drawable/btn_todo_active_sv"
-                    android:contentDescription="투두 버튼" />
-
-                <ImageView
-                    android:id="@+id/btn_wish"
-                    android:layout_width="60dp"
-                    android:layout_height="28dp"
-                    android:src="@drawable/btn_wish_deactive_sv"
-                    android:contentDescription="위시 버튼" />
-            </LinearLayout>
-
-            <FrameLayout
-                android:id="@+id/register_fragment_container"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="16dp" />
-        </LinearLayout>
-    </androidx.core.widget.NestedScrollView>
-</FrameLayout>
